Fix segfault caused by concurrent access to Channel Outputs
When stopping a sequence, the main channel output thread was still
running and calling the Channel Outputs' PrepData() and SendData()
functions, but on another thread, Sequence.cpp was calling
PrepData() and SendData() itself which was causing segfaults due
to two instances of PrepData() or SendData() from the same
Channel Output running at the same time.

Rather than having Sequence.cpp send the blanking data, check to
see if the Channel Output thread is running and kick it instead so
we aren't prepping and sending from two threads at the same time.