Add an Auto Layout button to the LED Panels config UI.
Panels are normally designed to be chained from left to right
with the arrows facing up.  When the Auto Layout button is used,
it will auto configure the panel orientation, output number, and
panel number as if the panels were laid out in this manner and
each panel row was driven by a separate output.