Update libfpp.so to less-restrictive LGPL and clarify 3rd-party sources
This commit updates libfpp.so and its source files to the less-
restrictive LGPL instead of the GPL.  This will allow plugin authors
to link against libfpp.so without requiring those plugins to be
licensed under the GPL themselves.

This commit also includes an updated LICENSE file and set of LICENSE.*
files to clarify which license FPP code is covered by as well as which
licenses are applicable to the various 3rd-Party files which are
included in the FPP repository.

The SunSet class used previously was incompatible with the LGPL and
has now been replaced with the SunRise class which is compatible.  This
change was required since these sunrise/sunset calculations are
performed inside libfpp.so.