Auto hide/show Restart/Reboot warning boxes
We now pass the restart and reboot flags back in the system status
API call.  This allows the restart/reboot warning boxes to be
automatically hidden or shown on a browser tab if they were
triggered by another action in another tab.  Make a change in one
tab and it should be reflected in the others.

Also a minor fix to set the new value for a setting in the settings
array if we set it on the server.