Security functionality, make FPP more IoT-law compliant
There are multiple IoT related laws now in effect or going into
effect regarding security of IoT devices, which FPP could be
considered to be a part of.

This commit adds new functionality to FPP to force the user to
decide whether to use the default security settings or enforce
stronger settings.  This is done by forcing the Initial Setup
page to show even if it has previously shown.

The user must choose to either leave the FPP web UI unprotected
or set a UI password.

The user must choose to either use the default fpp user OS
password of 'falcon' or choose a custom password.  This setting
is also on the Settings->System tab as well.

The user has the option of uploading SSH keys right on the
Initial Setup page (and Settings->System page later)