Allow each output in a strings config to supply its own UI label
The main use for this is to allow custom configs for a cape which
would skip some connected pins such as on the BBGW w/ WiFi enabled
without having to resort to using a group to set the correct port
number label for each individual pin.  This could also be used on a

Usage is to just add a "label": "value" entry to the desired output
array entries for items that need a custom label rather than using
a default label based on their numeric position within the output
array.