Standardize WiFi level to use Link Quality normalized to 0-100 range
Some WiFi adapters report their signal value in dBm, while some report a
percentage  There are a bunch that report a percentage that seems to be
incorrectly calculated.  The Link Quality on all of these appears to be
more consistently calculated, so use that instead.

Prior to FPP v6, the cards that report a percentage were all listed as
'excellent' signal in the UI because their positive percentage was higher
than the 'excellent' dBm threshold.  This should fix that issue, but will
not fix the perceive notion that FPP v6.x has lower WiFi performance or
quality than FPP v5.x had.