Change FPP Command Editor to use jQuery colpick
The FPP Command Editor page was using relying on the browsers color swatch implementation when allowing users to set a colour.
This works but isn't as flexible as the colour picker on the testing page which is using jQuery colpick which allows a nice color swatch and input of RGB/HEX values

I made the colour picker work on the existing input fields as to not upset any downstream code that is looks for these fields.
It also appears the colour picker will not activate correctly if the content is not visible on the screen, a small loop

Closes #1483