Try offset or locate the color picker better
calculate a negative top margin based off the modal/dialogs effective offset on the screen (content of the dialog is offset down by the dialogs top padding, the header, then the contents top padding.

Append colour picker to the modal-header (sounds weird I know), but this allows it ro pop over the modal's footer instead of being hidden under it (because it's part of the content)

Cancel interval loop & restart if function is called again.