Change hostname link to go to main status page.
On mobile devices without the FPP or Falcon logo, the hostname is
shown in the top left.  Make the hostname link go to the status page
just like the FPP and Falcon logo links do on desktop.

On mobile, this makes going to the status page 1-click instead of 3
to get there via the menu.

Fixes #1667