Fix a bug when using holiday names for scheduling.
When the holiday passed, the schedule broke because the start
date would be set to the next occurrence of the holiday.

This removes the offending code added in 557a9d4fdf and reworks
the UI to not need the full year when listing holiday names in
the schedule preview.

Fixes #2077