Fix Internet Health check
As per https://github.com/fail2ban/fail2ban/issues/313, example.com has moved
and the old ip address stopped responding.  Changed to use the ip address of
fppstatus because that same server is also used for checking for available updates.  Has to be an IP address because DNS may not be configured.

Also add a backup server to check (kulplights.com) but only if the first
server doesn't respond.

Closes #2176