ColorLight changes to support v13.x firmware
Channel Output now detects firmware on receiver cards by
sending discover packets and parsing the response(s) received.
The code still has an issue if both an older firmware card and
newer firmware card are chained together, one will need to be
upgraded or downgraded to work properly.  Additional research
and mods may be needed to get the same packets to work with
both firmware versions at the same time.

A simple wireshark packet dissector for ColorLight packets is also
included in the docs directory.  This will need to be updated as
other packet types are reverse-engineered.