Add Advanced-level setting to force Colorlight FW version
Default 'Auto' uses a discovery packet to detect the firmware
version on the reciever cards.  Setting can also be set to
"v2.x - v12.x" or "v13.x and up" to handle cases where the
discovery does not work for some reason, for istance if the
receivers are powered off when fppd is restarted.