Update how start channel is assigned on Effect Start command.
Since we can't have conditional values in the input UI for
FPP Commands, rather than setting the default to 1, leave it blank
as before and update the backend fppd code so that if a model is
selected, the code acts as if the start channel had been set to one
if it was left empty.  This should fix the original issue and
allow the code to still work as it did previously.

References #1596.
Fixes #2041